refactor: Change file structure, break play/pause functions out into utils
