+++ tests/HoverVideoPlayer.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, fireEvent, act } from '@testing-library/react';
 
-import { getVideoState, VIDEO_STATE } from '../src/utils';
+import { getVideoState, VIDEO_STATE } from '../src/utils/video';
 import HoverVideoPlayer from '../src';
 
 /**
